Silence avertissements.

+++ coeur/decoupage/analyseuse.cc
@@ -104,8 +104,8 @@ bool Analyseuse::sont_2_identifiants(int id1, int id2)
 	m_tableau_identifiant[id2]++;
 #endif
 
-	return m_identifiants[m_position].identifiant == id1
-			&& m_identifiants[m_position + 1].identifiant == id2;
+	return m_identifiants[m_position].identifiant == static_cast<size_t>(id1)
+			&& m_identifiants[m_position + 1].identifiant == static_cast<size_t>(id2);
 }
 
 bool Analyseuse::sont_3_identifiants(int id1, int id2, int id3)
@@ -120,9 +120,9 @@ bool Analyseuse::sont_3_identifiants(int id1, int id2, int id3)
 	m_tableau_identifiant[id3]++;
 #endif
 
-	return m_identifiants[m_position].identifiant == id1
-			&& m_identifiants[m_position + 1].identifiant == id2
-			&& m_identifiants[m_position + 2].identifiant == id3;
+	return m_identifiants[m_position].identifiant == static_cast<size_t>(id1)
+			&& m_identifiants[m_position + 1].identifiant == static_cast<size_t>(id2)
+			&& m_identifiants[m_position + 2].identifiant == static_cast<size_t>(id3);
 }
 
 int Analyseuse::identifiant_courant() const
@@ -131,7 +131,7 @@ int Analyseuse::identifiant_courant() const
 		return ID_INCONNU;
 	}
 
-	return m_identifiants[m_position].identifiant;
+	return static_cast<int>(m_identifiants[m_position].identifiant);
 }
 
 void Analyseuse::lance_erreur(const std::string &quoi, int type)
